requirements.txt
requests
feedparser
google-generativeai
python-dateutil